AddNewChanges
