--- erpnext/accounts/doctype/bank_transaction_accounting_entry/bank_transaction_accounting_entry.json
@@ -1,81 +1,0 @@
-{
- "autoname": "naming_series:",
- "creation": "2021-09-20 16:05:07.287559",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "naming_series",
-  "bank_transaction",
-  "detail",
-  "total_debit",
-  "total_credit",
-  "amended_from"
- ],
- "fields": [
-  {
-   "fieldname": "bank_transaction",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Bank Transaction",
-   "options": "Bank Transactions",
-   "reqd": 1
-  },
-  {
-   "fieldname": "detail",
-   "fieldtype": "Table",
-   "label": "Detail",
-   "options": "Bank Transactions Detail"
-  },
-  {
-   "fieldname": "naming_series",
-   "fieldtype": "Select",
-   "label": "Serie",
-   "options": "BAN-TRAN-.YYYY.-"
-  },
-  {
-   "fieldname": "amended_from",
-   "fieldtype": "Link",
-   "label": "Amended From",
-   "no_copy": 1,
-   "options": "Bank Transaction Accounting Entry",
-   "print_hide": 1,
-   "read_only": 1
-  },
-  {
-   "fieldname": "total_debit",
-   "fieldtype": "Currency",
-   "label": "Amount Debit",
-   "read_only": 1
-  },
-  {
-   "fieldname": "total_credit",
-   "fieldtype": "Currency",
-   "label": "Amount Credit",
-   "read_only": 1
-  }
- ],
- "is_submittable": 1,
- "modified": "2021-09-24 12:15:14.625649",
- "modified_by": "Administrator",
- "module": "Accounts",
- "name": "Bank Transaction Accounting Entry",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "track_changes": 1
-}
