--- erpnext/accounts/doctype/bank_transaction_accounting_entry/test_bank_transaction_accounting_entry.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2021, Frappe Technologies Pvt. Ltd. and Contributors
-# See license.txt
-from __future__ import unicode_literals
-
-# import frappe
-import unittest
-
-class TestBankTransactionAccountingEntry(unittest.TestCase):
-	pass
