+++ erpnext/accounts/doctype/journal_entry_account/journal_entry_account.json
@@ -50,7 +50,6 @@
    "oldfieldtype": "Link",
    "options": "Account",
    "print_width": "250px",
-   "reqd": 1,
    "search_index": 1,
    "width": "250px"
   },
@@ -272,7 +271,7 @@
  ],
  "idx": 1,
  "istable": 1,
- "modified": "2019-10-02 12:23:21.693443",
+ "modified": "2021-09-30 08:44:47.424731",
  "modified_by": "Administrator",
  "module": "Accounts",
  "name": "Journal Entry Account",
